backend/requirements.txt
# Core FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0

# HTTP client and security
aiohttp>=3.8.0
cryptography>=3.4.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]
python-multipart==0.0.6

# Data validation (COMPATIBLE Python 3.12)
pydantic==2.5.0
email-validator==2.1.0
pydantic-settings==2.1.0

# Environment
python-dotenv==1.0.0